weather-analytics-pipeline/requirements.txt
# ============================================================
# Python dependencies for the weather analytics pipeline
# ============================================================
# Install with: pip install -r requirements.txt
# ============================================================

# --- Data extraction ---
requests==2.31.0          # HTTP library for calling the Open-Meteo API
python-dotenv==1.0.1      # Loads environment variables from .env file

# --- Data warehouse ---
duckdb==1.1.0             # Embedded analytical database (our local warehouse)

# --- dbt (transformation layer) ---
dbt-core==1.8.0           # dbt's core engine
dbt-duckdb==1.8.0         # dbt adapter for DuckDB/MotherDuck

# --- Utilities ---
pandas==2.2.0             # DataFrame library (used in extraction for convenience)

